Lots of changes, see full commit msg;
`approaches.py`:
- Added profiling code to measure the time taken by the subprocesses
- Identified a small area where we needlessly recompute the classifier
  accuracy, removed that to gain about 2% speed gain (0.2s in 11.2s
  run). Profiled the program multiple times to ensure this is not a
  mistaken gain.

`data/`:
This folder contains data from runs/json objects etc. Anything that will
basically take a long time to generate, but can be re-used if stored.

Removed lots of the debugging print statements.

Profiling identified that `time.sleep(2)` was actually a blocking call,
and that this had increased the run time of the program by an order of
magnitude. Removed this for lots of speedup.